promote sx/sy from int16_t to int32_t
Screen coords from extreme frustum-side-plane straddlers can
exceed int16_t range (e.g. sx=55491 wraps to -10045), placing
vertices on the wrong side of the screen and creating giant
degenerate triangles that cover the viewport.

Changes:
  ScreenVertex.sx/sy       int16_t → int32_t
  fill_triangle_gouraud    x/y     int16_t → int32_t
  fill_triangle_flat       x/y     int16_t → int32_t
  draw_line                x/y     int16_t → int32_t
  transform_and_project    sx/sy   no longer cast to int16_t
  clip_edge_near           sx/sy   no longer cast to int16_t

sz stays int16_t (clamped to [-32767, 32767] via ndc_z * 32767).

+++ include/esp32gfx/mesh.hpp
@@ -14,7 +14,7 @@ struct Vertex {
 };
 
 struct ScreenVertex {
-    int16_t sx, sy;
+    int32_t sx, sy;
     int16_t sz;
     Color color;
     float clip_x, clip_y, clip_z, clip_w;
